requirements.txt
# NousData-Lab — Dependencias
# Framework de acceso a datos multi-formato

# ── Web Framework (API REST) ──────────────────────────
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
Werkzeug>=2.3.0

# ── Base de datos ─────────────────────────────────────
SQLAlchemy>=2.0.0

# ── Autenticación ─────────────────────────────────────
PyJWT>=2.8.0

# ── Formatos de datos ─────────────────────────────────
lxml>=4.9.0

# ── Utilidades ────────────────────────────────────────
python-dateutil>=2.8.2

# ── Desarrollo y testing (opcional) ───────────────────
# pytest>=7.4.0
# pytest-cov>=4.1.0
jsonschema>=4.19.0